Implement library
The base SDK has now been split up into an auth library and a utility
package for traverser and so on. This is based heavily on the
implementation found in the Base SDK with changes to initialization.

The major change is that you now provide credentials through a config
object rather than passing an API instance along.