Align terminology with Developer Portal (#4)
It's a bad idea to include user emails in JSON Web Tokens. The primary
concern is that our users will have their real emails exposed if any JWT
is ever inspected by anyone. There are a number of ways to avoid this,
and the easiest one is probably to just use opaque tokens instead of
emails.

Our developer portal will be issuing JWT credentials with opaque tokens,
so it makes little sense to have our SDK use terminology that matches
old and undesired behavior. Credentials issued with an email as issuer
should also make sense with this new terminology. "Issuer" could
describe an ID as well as an email, which should result in low friction
for "old" users reading the new documentation (when it's fully updated).

This commit is essentially a search-and-replace: "email" → "issuer"

It's not a breaking change as "email" still works. If both "email" and
"issuer" are present, "issuer" will be used. We'll just not mention the
support for "email" anywhere as it's considered deprecated.