Fix dependency vulnerabilities (#6)
The version of axios we're depending on has a known security
vulnerability, and since we've locked its version, NPM will dutifully
install the vulnerable version for our users. That's obviously not
acceptable.

This commit updates to the latest version, but also relaxes the version
lock to allow PATCH and MINOR version changes. If axios care about
semver, this should be fine and alleviate future headache.